// src/lib/stores/ui/types/modalType.ts
export type MODAL_VIEWS =
  | "INIT_VIEW"
  | "CONFIRM_VIEW"
  | "PIN_EDIT_VIEW"
  | "LOG_RECORD_DONE_VIEW"
  | "PROFILE_EDIT_VIEW"
  | "MATE_CREATE_MAP_VIEW"
  | "MATE_LOCATION_MAP_VIEW"
  | "DONE_VIEW"
  | "ANIMATION_ALERT_VIEW"
  | "PIN_DETAIL_MODAL_VIEW"
  | "DEPLOY_ALERT_VIEW"
  | "MATE_PARTICIPANT_USER_VIEW"
  | "MATE_REQUEST_VIEW"
  | "ACCESS_LOGIN_VIEW";